minidb: refactor: organize code into more modules
