+++ minidb/src/lib.rs
@@ -93,287 +93,33 @@
 #![warn(clippy::pedantic, missing_docs)]
 #![allow(clippy::doc_markdown)]
 
+mod builder;
 mod encryption;
+mod model;
 mod testing;
 
+pub use crate::builder::{KeySource, MiniDBBuilder};
+pub use crate::model::TableModel;
 #[cfg(feature = "macros")]
 pub use minidb_macros::Table;
 pub use redb;
 
 use std::path::PathBuf;
 
-use crate::encryption::{decrypt_bytes, derive_key_from_password, encrypt_bytes};
+use crate::encryption::{decrypt_bytes, encrypt_bytes};
 use anyhow::{Context, Result, anyhow};
 use argon2::password_hash::{SaltString, rand_core::OsRng};
-use chacha20poly1305::{KeyInit, XChaCha20Poly1305};
-use redb::{
-    Database, ReadableDatabase, ReadableTable, ReadableTableMetadata, TableDefinition,
-    WriteTransaction,
-};
+use chacha20poly1305::XChaCha20Poly1305;
+use redb::{Database, ReadableDatabase, ReadableTable, ReadableTableMetadata, TableDefinition};
 use serde::{Deserialize, Serialize};
 
-const META_TABLE: TableDefinition<&'static str, &[u8]> = TableDefinition::new("meta");
-const SETTINGS_TABLE: TableDefinition<&'static str, &[u8]> = TableDefinition::new("settings");
+pub(crate) const META_TABLE: TableDefinition<&'static str, &[u8]> = TableDefinition::new("meta");
+pub(crate) const SETTINGS_TABLE: TableDefinition<&'static str, &[u8]> =
+    TableDefinition::new("settings");
 
 const META_KEY_SALT: &str = "salt";
 
-type Initializer = Box<dyn Fn(&WriteTransaction) -> Result<()>>;
-type ArgonKey = [u8; 32];
-
-/// A table model. A table model is a struct that implements the [`TableModel`] trait.
-///
-/// ## Example
-///
-/// ```rust,no_run
-/// use minidb::TableModel;
-/// use serde::{Deserialize, Serialize};
-///
-/// #[derive(Serialize, Deserialize)]
-/// struct Person {
-///     id: String,
-///     name: String,
-///     age: u8,
-/// }
-///
-/// impl TableModel for Person {
-///     const TABLE: redb::TableDefinition<'_, &'static str, &[u8]> = redb::TableDefinition::new("people");
-///
-///     fn get_id(&self) -> &str {
-///         &self.id
-///     }
-///
-///     fn set_id(&mut self, id: String) {
-///         self.id = id;
-///     }
-/// }
-/// ```
-pub trait TableModel: Serialize + for<'de> Deserialize<'de> {
-    /// The table definition
-    const TABLE: TableDefinition<'_, &'static str, &[u8]>;
-
-    /// Returns the id of the table model
-    fn get_id(&self) -> &str;
-
-    /// Sets the id of the table model
-    fn set_id(&mut self, id: String);
-}
-
-/// A builder for a [`MiniDB`]
-///
-/// ## Example
-///
-/// ```rust,ignore
-/// use minidb::MiniDBBuilder;
-///
-/// // create the database with the path to the database file. The file can have any extension
-/// // but it's recommended to use `.redb` so you can differentiate
-/// // between a MiniDB/redb database and a SQLite/other embedded database
-/// let db = MiniDBBuilder::new("test.redb")
-///     .table::<Person>() // you must register all table models
-///     .table::<Car>()
-///     .build()
-///     .unwrap();
-/// ```
-pub struct MiniDBBuilder {
-    path: PathBuf,
-    initializers: Vec<Initializer>,
-    key_source: Option<KeySource>,
-}
-
-impl MiniDBBuilder {
-    /// Creates a new [`MiniDBBuilder`]
-    ///
-    /// ## Arguments
-    ///
-    /// * `path` - The path to the database file
-    ///
-    /// ## Returns
-    ///
-    /// A new [`MiniDBBuilder`]
-    ///
-    /// ## Example
-    ///
-    /// ```rust,no_run
-    /// use minidb::MiniDB;
-    ///
-    /// // create a MiniDB builder with the file path
-    /// let db = MiniDB::builder("test.redb");
-    /// ```
-    pub fn new<P>(path: P) -> Self
-    where
-        P: Into<PathBuf>,
-    {
-        Self {
-            path: path.into(),
-            initializers: Vec::new(),
-            key_source: None,
-        }
-    }
-
-    /// Registers a table model
-    ///
-    /// ## Arguments
-    ///
-    /// * `T` - The table model to register
-    ///
-    /// ## Returns
-    ///
-    /// A new [`MiniDBBuilder`]
-    ///
-    /// ## Example
-    ///
-    /// ```rust,ignore
-    /// use minidb::{MiniDB, Table};
-    ///
-    /// #[derive(Table)]
-    /// struct Person{
-    ///     #[key]
-    ///     id: String,
-    /// }
-    ///
-    /// // create a MiniDB builder with the file path
-    /// // and register the table Person
-    /// let db = MiniDB::builder("test.redb")
-    ///     .table::<Person>();
-    /// ```
-    #[must_use]
-    pub fn table<T>(mut self) -> Self
-    where
-        T: TableModel + 'static,
-    {
-        self.initializers.push(Box::new(|txn| {
-            txn.open_table(T::TABLE)
-                .map(|_| ())
-                .context("failed to init table")?;
-            Ok(())
-        }));
-        self
-    }
-
-    /// Sets the key source
-    ///
-    /// ## Arguments
-    ///
-    /// * `source` - The key source, can be a password, a pre-derived key, or a function that returns a key (`[u8; 32]`)
-    ///
-    /// ## Returns
-    ///
-    /// A new [`MiniDBBuilder`]
-    ///
-    /// ## Example
-    ///
-    /// ```rust,no_run
-    /// use minidb::{KeySource, MiniDB};
-    ///
-    /// // create a MiniDB builder with a password
-    /// let db = MiniDB::builder("test.redb")
-    ///     // skipping table registering for convenience
-    ///     .key_source(KeySource::Password("secretpassword".to_string()));
-    ///
-    /// // create a MiniDB builder with a pre-derived key
-    /// let key = [1u8; 32];
-    /// let db = MiniDB::builder("test.redb")
-    ///     // skipping table registering for convenience
-    ///     .key_source(KeySource::PreDerived(key));
-    ///
-    /// // create a MiniDB builder with a function that returns a key
-    /// fn key_provider() -> [u8; 32] {
-    ///     [1u8; 32]
-    /// }
-    ///
-    /// let db = MiniDB::builder("test.redb")
-    ///     // skipping table registering for convenience
-    ///     .key_source(KeySource::ExternalKeyProvider(Box::new(key_provider)));
-    /// ```
-    #[must_use]
-    pub fn key_source(mut self, source: KeySource) -> Self {
-        self.key_source = Some(source);
-        self
-    }
-
-    /// Builds the [`MiniDB`] from the builder
-    ///
-    /// ## Returns
-    ///
-    /// A new [`MiniDB`]
-    ///
-    /// ## Errors
-    ///
-    /// Returns an error if the database file already exists, if the bootstrap transaction fails, or if the key derivation fails
-    ///
-    /// ## Example
-    ///
-    /// ```rust,no_run
-    /// use minidb::{KeySource, MiniDB};
-    ///
-    /// // create a MiniDB
-    /// let db = MiniDB::builder("test.redb")
-    ///     // skipping table registering for convenience
-    ///     .key_source(KeySource::Password("secretpassword".to_string())) // if you want the database to be encrypted
-    ///     .build()
-    ///     .unwrap();
-    /// ```
-    pub fn build(self) -> Result<MiniDB> {
-        let db = Database::builder()
-            .create(&self.path)
-            .context("failed to create database file")?;
-
-        let txn = db.begin_write().context("failed to begin bootstrap txn")?;
-        {
-            let _ = txn
-                .open_table(META_TABLE)
-                .context("failed to init meta table")?;
-            let _ = txn
-                .open_table(SETTINGS_TABLE)
-                .context("failed to init settings table")?;
-        }
-        for init in self.initializers {
-            init(&txn)?;
-        }
-        txn.commit().context("failed to commit bootstrap")?;
-
-        let mut store = MiniDB { db, cipher: None };
-
-        if let Some(source) = self.key_source {
-            let key = match source {
-                KeySource::Password(pass) => {
-                    let salt = store
-                        .get_salt()
-                        .context("failed to get salt from meta table")?;
-
-                    derive_key_from_password(&pass, Some(salt), None)
-                        .context("failed to derive key")?
-                }
-
-                KeySource::PreDerived(key) => key,
-
-                KeySource::ExternalKeyProvider(provider_fn) => provider_fn(),
-            };
-
-            store.cipher = Some(XChaCha20Poly1305::new(&key.into()));
-        }
-
-        Ok(store)
-    }
-}
-
-/// The key source
-///
-/// ## Variants
-///
-/// * `KeySource::Password(String)` - The key source is a password
-/// * `KeySource::PreDerived(ArgonKey)` - The key source is a pre-derived key (`[u8; 32]`)
-/// * `KeySource::ExternalKeyProvider(Box<dyn Fn() -> ArgonKey>)` - The key source is a function that returns a key (`[u8; 32]`)
-pub enum KeySource {
-    /// The key source is a password
-    Password(String),
-
-    /// The key source is a pre-derived key (`[u8; 32]`)
-    PreDerived(ArgonKey),
-
-    /// The key source is a function that returns a key (`[u8; 32]`)
-    ExternalKeyProvider(Box<dyn Fn() -> ArgonKey>),
-}
+pub(crate) type ArgonKey = [u8; 32];
 
 /// A MiniDB
 ///
@@ -407,6 +153,14 @@ impl MiniDB {
         P: Into<PathBuf>,
     {
         MiniDBBuilder::new(path)
+    }
+
+    pub(crate) fn new(db: Database) -> Self {
+        Self { db, cipher: None }
+    }
+
+    pub(crate) fn set_cipher(&mut self, cipher: XChaCha20Poly1305) {
+        self.cipher = Some(cipher);
     }
 
     // EMD OF BUILDERS
@@ -488,7 +242,7 @@ impl MiniDB {
     /// Creates the table if it doesn't exist
     ///
     /// Use [`MiniDB::create_table`] instead
-    fn create_table_impl<K, V>(&self, table: TableDefinition<K, V>) -> Result<()>
+    pub(crate) fn create_table_impl<K, V>(&self, table: TableDefinition<K, V>) -> Result<()>
     where
         K: redb::Key,
         V: redb::Value,
@@ -737,7 +491,7 @@ impl MiniDB {
     }
 
     /// Retrieves the salt from the meta table
-    fn get_salt(&self) -> Result<String> {
+    pub(crate) fn get_salt(&self) -> Result<String> {
         let value: Option<String> = self
             .get_meta(META_KEY_SALT)
             .context("failed to get salt from meta table")?;
@@ -753,7 +507,7 @@ impl MiniDB {
     }
 
     /// Retrieves an item from the meta table
-    fn get_meta<T>(&self, key: &str) -> Result<Option<T>>
+    pub(crate) fn get_meta<T>(&self, key: &str) -> Result<Option<T>>
     where
         T: for<'a> Deserialize<'a>,
     {
@@ -821,7 +575,7 @@ impl MiniDB {
     }
 
     /// Sets an item in the meta table
-    fn set_meta<T>(&self, key: &str, value: &T) -> Result<()>
+    pub(crate) fn set_meta<T>(&self, key: &str, value: &T) -> Result<()>
     where
         T: Serialize,
     {
